+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - abstract time series support
 - sync with core uniform «objects»
 - multistage build
+- time series cluster support
 
 # 2.2.1 (2019-12-17)
 ## add